+++ packages/core/src/utils/getNavigationRouteForSelectLanguageScreen.ts
@@ -1,0 +1,6 @@
+import { CoreStackParamListBase } from "../types";
+
+export const getNavigationRouteForLanguageSelectScreen = (isOnboarding: boolean): keyof CoreStackParamListBase => {
+    if (isOnboarding) return "Root";
+    return "Settings"
+}